+++ eemeter/caltrack/hourly.py
@@ -203,14 +203,22 @@ class CalTRACKHourlyModel(SegmentedModel):
     occupancy_lookup : :any:`pandas.DataFrame`
         A dataframe with occupancy flags for each hour of the week and each segment.
         Segment names are columns, occupancy flags are 0 or 1.
-    temperature_bins : :any:`pandas.DataFrame`
+    occupied_temperature_bins : :any:`pandas.DataFrame`
         A dataframe of bin endpoint flags for each segment. Segment names are columns.
+    unoccupied_temperature_bins : :any:`pandas.DataFrame`
+        Ditto for the unoccupied mode.
     """
 
-    def __init__(self, segment_models, occupancy_lookup, temperature_bins):
-
+    def __init__(
+        self,
+        segment_models,
+        occupancy_lookup,
+        occupied_temperature_bins,
+        unoccupied_temperature_bins,
+    ):
         self.occupancy_lookup = occupancy_lookup
-        self.temperature_bins = temperature_bins
+        self.occupied_temperature_bins = occupied_temperature_bins
+        self.unoccupied_temperature_bins = unoccupied_temperature_bins
         super(CalTRACKHourlyModel, self).__init__(
             segment_models=segment_models,
             prediction_segment_type="one_month",
@@ -231,7 +239,8 @@ class CalTRACKHourlyModel(SegmentedModel):
             prediction_feature_processor=caltrack_hourly_prediction_feature_processor,
             prediction_feature_processor_kwargs={
                 "occupancy_lookup": self.occupancy_lookup,
-                "temperature_bins": self.temperature_bins,
+                "occupied_temperature_bins": self.occupied_temperature_bins,
+                "unoccupied_temperature_bins": self.unoccupied_temperature_bins,
             },
         )
 
@@ -245,7 +254,12 @@ class CalTRACKHourlyModel(SegmentedModel):
         data.update(
             {
                 "occupancy_lookup": self.occupancy_lookup.to_json(orient="split"),
-                "temperature_bins": self.temperature_bins.to_json(orient="split"),
+                "occupied_temperature_bins": self.occupied_temperature_bins.to_json(
+                    orient="split"
+                ),
+                "unoccupied_temperature_bins": self.unoccupied_temperature_bins.to_json(
+                    orient="split"
+                ),
             }
         )
         return data
@@ -268,14 +282,19 @@ class CalTRACKHourlyModel(SegmentedModel):
         c = cls(
             segment_models,
             occupancy_lookup,
-            pd.read_json(data.get("temperature_bins"), orient="split"),
+            pd.read_json(data.get("occupied_temperature_bins"), orient="split"),
+            pd.read_json(data.get("unoccupied_temperature_bins"), orient="split"),
         )
 
         return c
 
 
 def caltrack_hourly_fit_feature_processor(
-    segment_name, segmented_data, occupancy_lookup, temperature_bins
+    segment_name,
+    segmented_data,
+    occupancy_lookup,
+    occupied_temperature_bins,
+    unoccupied_temperature_bins,
 ):
     """A function that takes in temperature data and returns a dataframe of
     features suitable for use with :any:`eemeter.fit_caltrack_hourly_model_segment`.
@@ -290,13 +309,16 @@ def caltrack_hourly_fit_feature_processor(
     occupancy_lookup : :any:`pandas.DataFrame`
         A dataframe with occupancy flags for each hour of the week and each segment.
         Segment names are columns, occupancy flags are 0 or 1.
-    temperature_bins : :any:`pandas.DataFrame`
+    occupied_temperature_bins : :any:`pandas.DataFrame`
         A dataframe of bin endpoint flags for each segment. Segment names are columns.
+    unoccupied_temperature_bins : :any:`pandas.DataFrame`
+        Ditto for the unoccupied mode.
 
     Returns
     -------
     features : :any:`pandas.DataFrame`
         A dataframe of features with the following columns:
+
         - 'meter_value': the observed meter value
         - 'hour_of_week': 0-167
         - 'bin_<0-6>_occupied': temp bin feature, or 0 if unoccupied
@@ -310,26 +332,56 @@ def caltrack_hourly_fit_feature_processor(
 
     # get temperature bin features
     temperatures = segmented_data.temperature_mean
-    bin_endpoints_list = (
-        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
+    occupied_bin_endpoints_list = (
+        occupied_temperature_bins[segment_name]
+        .index[occupied_temperature_bins[segment_name]]
+        .tolist()
     )
-    temperature_bin_features = compute_temperature_bin_features(
-        segmented_data.temperature_mean, bin_endpoints_list
+    unoccupied_bin_endpoints_list = (
+        unoccupied_temperature_bins[segment_name]
+        .index[unoccupied_temperature_bins[segment_name]]
+        .tolist()
+    )
+    occupied_temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, occupied_bin_endpoints_list
+    )
+    occupied_temperature_bin_features[occupancy_feature == 0] = 0
+    occupied_temperature_bin_features.rename(
+        columns={
+            c: "{}_occupied".format(c)
+            for c in occupied_temperature_bin_features.columns
+        },
+        inplace=True,
+    )
+    unoccupied_temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, unoccupied_bin_endpoints_list
+    )
+    unoccupied_temperature_bin_features[occupancy_feature == 1] = 0
+    unoccupied_temperature_bin_features.rename(
+        columns={
+            c: "{}_unoccupied".format(c)
+            for c in unoccupied_temperature_bin_features.columns
+        },
+        inplace=True,
     )
 
     # combine features
     return merge_features(
         [
             segmented_data[["meter_value", "hour_of_week"]],
-            occupancy_feature,
-            temperature_bin_features,
+            occupied_temperature_bin_features,
+            unoccupied_temperature_bin_features,
             segmented_data.weight,
         ]
     )
 
 
 def caltrack_hourly_prediction_feature_processor(
-    segment_name, segmented_data, occupancy_lookup, temperature_bins
+    segment_name,
+    segmented_data,
+    occupancy_lookup,
+    occupied_temperature_bins,
+    unoccupied_temperature_bins,
 ):
     """A function that takes in temperature data and returns a dataframe of
     features suitable for use inside :any:`eemeter.CalTRACKHourlyModel`.
@@ -344,13 +396,16 @@ def caltrack_hourly_prediction_feature_processor(
     occupancy_lookup : :any:`pandas.DataFrame`
         A dataframe with occupancy flags for each hour of the week and each segment.
         Segment names are columns, occupancy flags are 0 or 1.
-    temperature_bins : :any:`pandas.DataFrame`
+    occupied_temperature_bins : :any:`pandas.DataFrame`
         A dataframe of bin endpoint flags for each segment. Segment names are columns.
+    unoccupied_temperature_bins : :any:`pandas.DataFrame`
+        Ditto for the unoccupied mode.
 
     Returns
     -------
     features : :any:`pandas.DataFrame`
         A dataframe of features with the following columns:
+
         - 'hour_of_week': 0-167
         - 'bin_<0-6>_occupied': temp bin feature, or 0 if unoccupied
         - 'bin_<0-6>_unoccupied': temp bin feature or 0 in occupied
@@ -369,19 +424,45 @@ def caltrack_hourly_prediction_feature_processor(
 
     # get temperature bin features
     temperatures = segmented_data
-    bin_endpoints_list = (
-        temperature_bins[segment_name].index[temperature_bins[segment_name]].tolist()
+    occupied_bin_endpoints_list = (
+        occupied_temperature_bins[segment_name]
+        .index[occupied_temperature_bins[segment_name]]
+        .tolist()
     )
-    temperature_bin_features = compute_temperature_bin_features(
-        segmented_data.temperature_mean, bin_endpoints_list
+    unoccupied_bin_endpoints_list = (
+        unoccupied_temperature_bins[segment_name]
+        .index[unoccupied_temperature_bins[segment_name]]
+        .tolist()
+    )
+    occupied_temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, occupied_bin_endpoints_list
+    )
+    occupied_temperature_bin_features[occupancy_feature == 0] = 0
+    occupied_temperature_bin_features.rename(
+        columns={
+            c: "{}_occupied".format(c)
+            for c in occupied_temperature_bin_features.columns
+        },
+        inplace=True,
+    )
+    unoccupied_temperature_bin_features = compute_temperature_bin_features(
+        segmented_data.temperature_mean, unoccupied_bin_endpoints_list
+    )
+    unoccupied_temperature_bin_features[occupancy_feature == 1] = 0
+    unoccupied_temperature_bin_features.rename(
+        columns={
+            c: "{}_unoccupied".format(c)
+            for c in unoccupied_temperature_bin_features.columns
+        },
+        inplace=True,
     )
 
     # combine features
     return merge_features(
         [
             hour_of_week_feature,
-            occupancy_feature,
-            temperature_bin_features,
+            occupied_temperature_bin_features,
+            unoccupied_temperature_bin_features,
             segmented_data.weight,
         ]
     )
@@ -404,25 +485,6 @@ def fit_caltrack_hourly_model_segment(segment_name, segment_data):
         A model that represents the fitted model.
     """
 
-    def _get_hourly_model_formula(data):
-        if (np.sum(data.loc[data.weight > 0].occupancy) == 0) or (
-            np.sum(data.loc[data.weight > 0].occupancy)
-            == len(data.loc[data.weight > 0].occupancy)
-        ):
-            bin_occupancy_interactions = "".join(
-                [" + {}".format(c) for c in data.columns if "bin" in c]
-            )
-            return "meter_value ~ C(hour_of_week) - 1{}".format(
-                bin_occupancy_interactions
-            )
-        else:
-            bin_occupancy_interactions = "".join(
-                [" + {}:C(occupancy)".format(c) for c in data.columns if "bin" in c]
-            )
-            return "meter_value ~ C(hour_of_week) - 1{}".format(
-                bin_occupancy_interactions
-            )
-
     warnings = []
     if segment_data.dropna().empty:
         model = None
@@ -438,7 +500,15 @@ def fit_caltrack_hourly_model_segment(segment_name, segment_data):
                 },
             )
         )
+
     else:
+
+        def _get_hourly_model_formula(data):
+            return "meter_value ~ C(hour_of_week) - 1{}".format(
+                "".join(
+                    [" + {}".format(c) for c in data.columns if c.startswith("bin")]
+                )
+            )
 
         formula = _get_hourly_model_formula(segment_data)
         model = smf.wls(formula=formula, data=segment_data, weights=segment_data.weight)
@@ -464,7 +534,10 @@ def fit_caltrack_hourly_model_segment(segment_name, segment_data):
 
 
 def fit_caltrack_hourly_model(
-    segmented_design_matrices, occupancy_lookup, temperature_bins
+    segmented_design_matrices,
+    occupancy_lookup,
+    occupied_temperature_bins,
+    unoccupied_temperature_bins,
 ):
     """Fit a CalTRACK hourly model
 
@@ -476,8 +549,10 @@ def fit_caltrack_hourly_model(
     occupancy_lookup : :any:`pandas.DataFrame`
         A dataframe with occupancy flags for each hour of the week and each segment.
         Segment names are columns, occupancy flags are 0 or 1.
-    temperature_bins : :any:`pandas.DataFrame`
+    occupied_temperature_bins : :any:`pandas.DataFrame`
         A dataframe of bin endpoint flags for each segment. Segment names are columns.
+    unoccupied_temperature_bins : :any:`pandas.DataFrame`
+        Ditto for the unoccupied mode.
 
     Returns
     -------
@@ -493,7 +568,13 @@ def fit_caltrack_hourly_model(
         for segment_model in segment_models
         for warning in segment_model.warnings
     ]
-    model = CalTRACKHourlyModel(segment_models, occupancy_lookup, temperature_bins)
+
+    model = CalTRACKHourlyModel(
+        segment_models,
+        occupancy_lookup,
+        occupied_temperature_bins,
+        unoccupied_temperature_bins,
+    )
 
     model_results = CalTRACKHourlyModelResults(
         status="SUCCEEDED",