+++ CHANGELOG.md
@@ -4,7 +4,12 @@ Changelog
 Development
 -----------
 
-* Placeholder
+* Remove python27 support.
+* Update Pipfile lock.
+* Update `fit_temperature_bins` to potentially take an `occupancy_lookup` in order to
+  fit different temperature bins for occupied/unoccupied modes.
+* Update CalTRACK hourly model formula to use different bins for occupied and
+  unoccupied mode.
 
 2.10.11
 -------